% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{annotateTerms}
\alias{annotateTerms}
\title{A function that takes a single term or a vector of terms and queries a specified ontology.}
\usage{
annotateTerms(term, ebiOntology, exact = FALSE, verbose = FALSE)
}
\arguments{
\item{term}{a single term or vector of terms}

\item{ebiOntology}{(ebi ontologies)[https://www.ebi.ac.uk/ols/ontologies]}

\item{exact}{Strings to be transformed to NO}

\item{verbose}{Transform all to upper case to make substitution easier}
}
\description{
A function that takes a single term or a vector of terms and queries a specified ontology.
}
\keyword{ontology}
