% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{standardizeYesNo}
\alias{standardizeYesNo}
\title{A function that standardizes binary columns of Y/N}
\usage{
standardizeYesNo(col, yes = NULL, no = NULL, upper = TRUE, verbose = TRUE)
}
\arguments{
\item{col}{Column or vector to be standardized}

\item{yes}{Strings to be transformed to YES}

\item{no}{Strings to be transformed to NO}

\item{upper}{Transform all to upper case to make substitution easier}

\item{verbose}{Output a message that identifies left over strings in the column that are not YES or NO}
}
\description{
This function allows you to input a vector and strings that should be transformed to
YES or NO and outputs a standardized column
}
\keyword{cleaning}
\keyword{data}
