% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDictionary.R
\name{valueSummary}
\alias{valueSummary}
\title{A function to output information about a column or to compare two columns}
\usage{
valueSummary(col_orig, col_new)
}
\arguments{
\item{col_orig}{Original column}

\item{col_new}{New column}
}
\description{
This function allows you to input two columns and outputs
simple descriptors of each column for easy comparison
}
\keyword{summary}
