% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDictionary.R
\name{buildDataDict}
\alias{buildDataDict}
\title{A function to build simple data dictionaries from a dataframe.}
\usage{
buildDataDict(df, IDcols = NULL)
}
\arguments{
\item{IDcols}{Are columns in this dataframe used for identification? Defaults to NULL.}
}
\description{
This function allows you to input a dataframe and outputs
simple descriptors of each field in that dataframe.
}
\keyword{data}
\keyword{dictionary}
